Remove C# scripts and provide Python equivalents
